split makernote logic to lib
